+++ src/duckherder_extension.cpp
@@ -2,35 +2,13 @@
 
 #include "duckdb.hpp"
 #include "duckherder_extension.hpp"
+#include "duckherder_functions.hpp"
 #include "duckherder_extension_instance_state.hpp"
-#include "duckherder_pragmas.hpp"
 #include "duckherder_storage.hpp"
-#include "query_history_query_function.hpp"
-#include "query_execution_stats_query_function.hpp"
-#include "server/driver/distributed_server_function.hpp"
 
 namespace duckdb {
 
 namespace {
-
-// Successful execution result.
-constexpr bool SUCCESS = true;
-
-// Get database instance from expression state.
-// Returned instance ownership lies in the given [`state`].
-DatabaseInstance &GetDatabaseInstance(ExpressionState &state) {
-	auto *executor = state.root.executor;
-	auto &client_context = executor->GetContext();
-	return *client_context.db.get();
-}
-
-// Clear stats for query recorder.
-void ClearQueryRecorderStats(const DataChunk &args, ExpressionState &state, Vector &result) {
-	auto &duckdb_instance = GetDatabaseInstance(state);
-	auto &instance_state = GetInstanceStateOrThrow(duckdb_instance);
-	instance_state.GetQueryRecorder()->ClearQueryRecords();
-	result.Reference(Value(SUCCESS));
-}
 
 void LoadInternal(ExtensionLoader &loader) {
 	auto &db = loader.GetDatabaseInstance();
@@ -40,31 +18,7 @@ void LoadInternal(ExtensionLoader &loader) {
 	// Set extension state.
 	SetInstanceState(db, make_shared_ptr<DuckherderInstanceState>());
 
-	// Register pragma functions to register and unregister remote table.
-	DuckherderPragmas::RegisterPragmas(loader);
-
-	// Register function to get query stats.
-	loader.RegisterFunction(GetQueryHistory());
-
-	// Register function to get query execution stats from driver.
-	loader.RegisterFunction(GetQueryExecutionStats());
-
-	// Register function to clear query recorder stats.
-	ScalarFunction clear_recorder_stats_function("duckherder_clear_query_recorder_stats",
-	                                             /*arguments=*/ {},
-	                                             /*return_type=*/LogicalType {LogicalTypeId::BOOLEAN},
-	                                             ClearQueryRecorderStats);
-	loader.RegisterFunction(clear_recorder_stats_function);
-
-	// Register distributed server control functions, which could be local usage.
-	loader.RegisterFunction(GetStartLocalServerFunction());
-	loader.RegisterFunction(GetStopLocalServerFunction());
-	loader.RegisterFunction(GetRegisterOrReplaceDriverFunction());
-
-	// Register worker management functions.
-	loader.RegisterFunction(GetWorkerCountFunction());
-	loader.RegisterFunction(GetRegisterWorkerFunction());
-	loader.RegisterFunction(GetStartStandaloneWorkerFunction());
+	RegisterDuckherderFunctions(loader);
 }
 
 } // namespace