+++ src/include/duckherder_functions.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace duckdb {
+
+class ExtensionLoader;
+
+void RegisterDuckherderFunctions(ExtensionLoader &loader);
+
+} // namespace duckdb